% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_obs.R
\docType{data}
\name{Q_obs}
\alias{Q_obs}
\title{Observed daily flows}
\format{
A object with 13514 rows and 1 variable:
\describe{
\item{Q_obs}{observations, runoff in mm per day}
}
}
\usage{
Q_obs
}
\description{
A column containing values of observed flow.
}
\keyword{datasets}
