% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_sim.R
\docType{data}
\name{Q_sim}
\alias{Q_sim}
\title{Simulated daily flows}
\format{
A object with 13514 rows and 1 variable:
\describe{
\item{Q_sim}{simulations, runoff in mm per day}
}
}
\usage{
Q_sim
}
\description{
A column containing values of simulated flow.
}
\keyword{datasets}
