% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroDCindex.R
\name{hydroDC_Index}
\alias{hydroDC_Index}
\title{Duration Curve Hydrological Model Index}
\usage{
hydroDC_Index(Q_obs, Q_sim)
}
\arguments{
\item{Q_obs}{Column with daily observed flows}

\item{Q_sim}{Column with daily simulated flows}
}
\description{
Duration Curve Hydrological Model Index
}
